frontend/requirement.txt
# Core Framework
Flask==3.1.2

# Database
psycopg2-binary==2.9.10

# Validation
email-validator==2.3.0

# Computer Vision
opencv-python==4.12.0.88

# Utilities
numpy==2.2.6
